fix #72: remove "root" names
Root names (i.e. artifacts with the whole name of `REQ`, `SPC`, etc)
were a hassle. They were not creatable or editable by the user
and were ALWAYS a partof every artifact with any name, which means
that they would reflect a "completed" and "tested" percentage --
but that percentage was not necessarily indicative of anything.

Instead of having required roots, the user should have the power
to specify what their root is. By convention the root is
`REQ-purpose` which is a clearer and simpler name.